Fix Expect header handling
Previously the Expect header would be ignored as an earlier read would
time out waiting for the content. "HTTP/1.1 100 Continue" needs to be
sent before the content is read.

Also, this commit makes the comparison for the header case insensitive.
From RFC 2616:

http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.20

"Comparison of expectation values is case-insensitive for unquoted
tokens (including the 100-continue token), and is case-sensitive for
quoted-string expectation-extensions."

From RFC 7231:

http://tools.ietf.org/html/rfc7231#section-5.1.1

"The Expect field-value is case-insensitive."